+++ PdfSharp.Xamarin.Forms/Renderers/PDFLabelRender.cs
@@ -1,4 +1,5 @@
 ﻿using PdfSharp.Xamarin.Forms.Attributes;
+using PdfSharp.Xamarin.Forms.Extensions;
 using PdfSharpCore.Drawing;
 using PdfSharpCore.Fonts;
 using Xamarin.Forms;
